Crate serenity_poise_macros

Procedural macros used in poise, like command

Attribute Macros

command

This macro transforms plain functions into poise bot commands.

Derive Macros

ChoiceParameter

Use this derive macro on an enum to easily generate a choice parameter type. A choice parameter is mainly useful in slash commands. It allows you to constrain input to a fixed set of choices.

Modal

See Modal trait documentation

SlashChoiceParameterDeprecated

See ChoiceParameter